backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
azure-ai-formrecognizer
azure-core
networkx
pydantic
# Optional (for local OCR if ever needed):
# pillow
# pytesseract
